Binary Tree Width
Added functionality to find the maximum width among all levels of a binary tree.

Algorithm using an adaptation of breadth first search, using a Queue data structure to iterate through the nodes.

Also changed the BinaryTree.__init__ function to initalise the tree from a list of nodes, rather than a dictionary as it is easier for a user to initalise this way.

TODO: modify the BinaryTree._strHelper function to use a similar algorithm to the width function.